Refactor OAStats pipeline
This is a significant refactoring of the pipeline in preparation for
moving to Postgres. Much of the code has been simplified and removed
since we no longer need to deal with adding a summary collection. Some
further optimizations have been made to push expensive operations as far
down the pipeline as possible. This allows for more aggressive filtering
up front that drastically cuts down on the work being performed per
request.

Date filtering is now done within the pipeline. This was done to
simplify the execution of the pipeline, which previously relied on grep
to feed requests into the pipeline.

The dogpile.cache package has been added to further improve performance
for operations that would benefit from caching.