Add output to CSV for PostGres COPY
This adds a CSV formatted output suitable for passing to the PostGres
COPY command. Only the requests table is populated in this way; the
underlying documents and identity tables get updated during the run.
This is necessary in order to deal with foreign keys, but the bulk of
the inserts will be from requests.